CB-1 Add elastic memory mapped and native memory Bytes

+++ lang8/src/main/java/net/openhft/chronicle/core/Memory.java
@@ -83,5 +83,5 @@ public interface Memory {
 
     boolean compareAndSwapLong(Object underlyingObject, long offset, long expected, long value);
 
-    long pageSize();
+    int pageSize();
 }
